requirements.txt
aiogram==3.26.0
fastapi==0.135.2
httpx==0.28.1
python-dotenv==1.2.2
uvicorn==0.42.0

